Move code to src folder & rename main
